+++ src/scene/shader-lib/chunks/gsplat/gsplatOutput.js
