+++ src/scene/shader-lib/chunks/gsplat/vert/gsplatOverrides.js
@@ -1,0 +1,15 @@
+export default /* glsl */`
+
+void centerOverride(inout vec3 center) {
+}
+
+void scaleOverride(in vec3 center, inout vec3 covA, inout vec3 covB) {
+}
+
+void alphaOverride(in vec3 center, inout float alpha) {
+}
+
+void colorOverride(in vec3 center, inout vec3 color) {
+}
+
+`;
